Remove test-commands from manifest
There was a bug in flatpak-builder's JSON schema for manifests that incorrectly required `test-commands` to be a string rather than a list of strings. This has been fixed but not yet rolled out into the `flatpak-builder` build used by the Flathub infrastructure.

But in this manifest the `test-commands` are unused, since the module also specifies `"run-tests": false`. Remove the offending commands.