Downgrade the runtime to GNOME 42
The Asyncio of clubhouse has compatibility issue with Python 3.10+ [1].
And, GNOME 43 runtime has Python 3.10. However, we has not had a way to
fix the issue, yet. Therefore, downgrade the runtime to GNOME 42 to gain
more time to fix the Asyncio compatibilty issue between clubhouse and
Python 3.10+.

Follow the commit ("Downgrade the runtime to GNOME 42") of upstream
clippy [2].

Besides, downgrade gst-plugins-bad to 1.18.6 for GNOME 42 runtime
compatiblity, too.

[1]: https://github.com/flathub/com.hack_computer.Clubhouse/issues/22#issuecomment-1352841349
[2]: https://github.com/endlessm/clippy